+++ core/autoload/backup_core.gd
@@ -7,11 +7,11 @@ signal backup_saved(was_auto: bool)
 signal backup_failed(was_auto: bool)
 
 func _ready() -> void:
+	get_window().close_requested.connect(_cleanup_backups)
 	Settings.define_preset("files", "auto_backup", true)
 	Settings.define_preset("files", "auto_backup_iterval_minutes", 1)
 	Settings.define_preset("files", "keep_backup_for_days", 10)
 	_handle_auto_save()
-	get_window().close_requested.connect(_cleanup_backups)
 
 
 # Cleanups backups