+++ core/scripts/backup_status.gd
@@ -9,11 +9,11 @@ func _ready() -> void:
 
 
 func _on_backup_saved(was_auto: bool) -> void:
-	show()
 	texture = load("res://assets/backup.png")
 	tooltip_text = "Backup Status\n{0} backup saved: {1}".format([
 		"Auto" if was_auto else "Manual", Time.get_datetime_string_from_system(false, true)
 	])
+	modulate = Color.WHITE
 	show()
 	hide_timer.start()
 
@@ -23,5 +23,6 @@ func _on_backup_failed(was_auto: bool) -> void:
 	tooltip_text = "Backup Status\n{0} backup failed: {1}".format([
 		"Auto" if was_auto else "Manual", Time.get_datetime_string_from_system(false, true)
 	])
+	modulate = Color.WHITE
 	show()
 	hide_timer.start()